+++ src/index.ts
@@ -1,4 +1,14 @@
+import _traverse from 'traverse'
 export { TraverseContext, Traverse } from 'traverse'
-const _traverse = require('traverse') //tslint:disable-line
 export const traverse = _traverse
-export * from './keepBranchesIncluding'
+export * from './lib/keepBranchesIncluding'
+
+export type Primitive = boolean | number | string | null | undefined
+
+export interface GenericObj {
+  [key: string]: any
+}
+
+export interface Hashed {
+  [key: string]: Primitive
+}